+++ joppy2/Presentation/Event/EventScreen.swift
@@ -19,7 +19,8 @@ struct EventScreen: View {
     @ScaledMetric(relativeTo: .body) var verticalPadding: CGFloat = 2
     @ScaledMetric(relativeTo: .body) var cornerRadiusValue: CGFloat = 8
     @State var eventFieldTypes: EventFieldTypes = .toC
-    
+    @State var isShowSheet: Bool = false
+
     var body: some View {
         NavigationStack {
             VStack {
@@ -36,11 +37,10 @@ struct EventScreen: View {
                 VStack {
                     switch eventFieldTypes {
                     case .toC:
-                        //参加用画面、カードで並べる
-                        Text("参加したい")
-                    case .toB:
-                        Text("開催したい")
                         HeldScreen()
+                    case .toB:
+                        EditButton(isShowSheet: $isShowSheet)
+
                     }
                 }
                 .animation(.easeInOut, value: eventFieldTypes)
